% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brewer_cols.R
\name{brewer_dhsc}
\alias{brewer_dhsc}
\title{DHSC colour palettes}
\usage{
brewer_dhsc(theme = "dhsc", n = Inf, names = FALSE)
}
\arguments{
\item{theme}{string; current option is only "ohid"}

\item{n}{number; number of colours required}

\item{names}{logical; TRUE will return a named character vector}
}
\description{
DHSC colour palettes
}
\details{
Returns a character vector containing hex colours based on the colour palette
requested
}
\examples{
library(ggplot2)
library(DHSCcolours)

categories <- c("Car", "Bike", "Horse", "Running")
df <- data.frame(category = categories,
                 value = runif(4))

pal <- brewer_dhsc(theme = "ohid",
                   n = 4)
p <- ggplot(df, aes(x = category, y = value)) +
        geom_col(aes(fill = category)) +
        labs(title = "Methods of transport") +
        theme_minimal() +
        scale_fill_manual(values = pal)
p

## OR

pal <- brewer_dhsc(theme = "ohid",
                   names = TRUE)
pal <- pal[c("DHSC teal", "red", "maroon", "purple")]
names(pal) <- categories
p <- ggplot(df, aes(x = category, y = value)) +
        geom_col(aes(fill = category)) +
        labs(title = "Methods of transport") +
        theme_minimal() +
        scale_fill_manual(values = pal)
p
}
