% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{DHSC_accessible_3}
\alias{DHSC_accessible_3}
\title{DHSC_accessible_3}
\usage{
DHSC_accessible_3()
}
\value{
A list of three colours for a colour blind friendly palette.
}
\description{
An accessible 3 colour palette.
}
