% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{DHSC_accessible_5}
\alias{DHSC_accessible_5}
\title{DHSC_accessible_5}
\usage{
DHSC_accessible_5()
}
\value{
A list of five colours for a colour blind friendly palette.
}
\description{
An accessible 5 colour palette.
}
