% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{DHSC_accessible_5_gradient}
\alias{DHSC_accessible_5_gradient}
\title{DHSC_accessible_5_gradient}
\usage{
DHSC_accessible_5_gradient()
}
\value{
A list of five colours for a colour blind friendly palette.
}
\description{
An accessible 5 colour palette for ordinal data.
}
