% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{DHSC_accessible_scales}
\alias{DHSC_accessible_scales}
\title{DHSC_accessible_scales}
\usage{
DHSC_accessible_scales(N, aesthetics = c("colour", "fill"))
}
\arguments{
\item{N}{The number of colours to use, either 3 or 4}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful,
for example, to apply colour settings to the colour and fill aesthetics
at the same time, via aesthetics = c("colour", "fill").}
}
\value{

}
\description{
This function encodes ggplot color scales which are colour blind
friendly. It currently includes a 3 and 4 colour option.
}
