% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{dhsc_dark_green}
\alias{dhsc_dark_green}
\title{dhsc_dark_green
The colour code for DHSC dark green.}
\usage{
dhsc_dark_green()
}
\value{

}
\description{
dhsc_dark_green
The colour code for DHSC dark green.
}
