% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{dhsc_light_green}
\alias{dhsc_light_green}
\title{dhsc_light_green
The colour code for DHSC light green.}
\usage{
dhsc_light_green()
}
\description{
dhsc_light_green
The colour code for DHSC light green.
}
