% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{dhsc_pal}
\alias{dhsc_pal}
\title{dhsc_pal}
\usage{
dhsc_pal(n, alpha = 1, begin = 0, end = 1, direction = 1)
}
\arguments{
\item{n}{The number of colors (\eqn{\ge 1}) to be in the palette.}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in [0,1] at which the viridis colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the viridis colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are ordered from darkest to lightest. If -1, the order of colors is reversed.}
}
\description{
This function creates a vector of n equally spaced colors along
the list of DHSC colours. Reordering the list will change the colours it
generates. The code is based on the virids function from viridislite.
}
\examples{
dhsc_pal(2)
}
