% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{dhsc_primary}
\alias{dhsc_primary}
\title{dhsc_primary
The colour code for the DHSC primary colour.}
\usage{
dhsc_primary()
}
\description{
dhsc_primary
The colour code for the DHSC primary colour.
}
