% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{dhsc_table}
\alias{dhsc_table}
\title{dhsc_table}
\usage{
dhsc_table(table_data, ...)
}
\arguments{
\item{table_data}{An R object, which is typically a matrix or data frame.}

\item{...}{Other arguments to pass to kable.}
}
\value{

}
\description{
For inserting consistent tables into rmarkdown documents. This
is just A wrapper for knitr::kable() in which the default alignment of
text within columns has been changed to be left-aligned.
}
