% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_themes.R
\name{html_ohid}
\alias{html_ohid}
\title{OHID-branded html documents}
\usage{
html_ohid(toc = FALSE, includes = NULL, ...)
}
\arguments{
\item{toc}{logical; whether to include a table of contents}

\item{includes}{Named list of additional content to include within the
document (typically created using the includes function from rmarkdown
package).}

\item{...}{other arguments to pass to html_document()}
}
\description{
Convert your markdown documents to OHID brand
}
\concept{document template}
