% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{scale_colour_dhsc_blue_green_c}
\alias{scale_colour_dhsc_blue_green_c}
\title{scale_colour_dhsc_blue_green}
\usage{
scale_colour_dhsc_blue_green_c()
}
\description{
Use a continuous colour scale from DHSC blue to green.
}
