% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{scale_colour_dhsc_blue_yellow}
\alias{scale_colour_dhsc_blue_yellow}
\title{scale_colour_dhsc_blue_yellow}
\usage{
scale_colour_dhsc_blue_yellow()
}
\description{
Use a continuous colour scale from DHSC blue to yellow
}
