% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{scale_colour_dhsc_c}
\alias{scale_colour_dhsc_c}
\title{scale_colour_dhsc_c
Use a continuous colour scale from DHSC green to purple.}
\usage{
scale_colour_dhsc_c()
}
\description{
scale_colour_dhsc_c
Use a continuous colour scale from DHSC green to purple.
}
