% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{scale_colour_dhsc_red_green}
\alias{scale_colour_dhsc_red_green}
\title{scale_colour_dhsc_red_green}
\usage{
scale_colour_dhsc_red_green()
}
\description{
Use a continuous colour scale from DHSC green to red.
}
