% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{scale_fill_dhsc_blue_yellow_c}
\alias{scale_fill_dhsc_blue_yellow_c}
\title{scale_fill_dhsc_blue_yellow}
\usage{
scale_fill_dhsc_blue_yellow_c()
}
\description{
Use a continuous fill scale from DHSC blue to yellow
}
