% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{scale_fill_dhsc_d}
\alias{scale_fill_dhsc_d}
\title{scale_fill_dhsc_d
Use the DHSC discrete fill scale.}
\usage{
scale_fill_dhsc_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  aesthetics = "fill"
)
}
\arguments{
\item{...}{Other arguments passed on to [discrete_scale()]}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{begin}{The (corrected) hue in [0,1] at which the viridis colormap begins.}

\item{end}{The (corrected) hue in [0,1] at which the viridis colormap ends.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors
are ordered from darkest to lightest. If -1, the order of colors is reversed.}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful,
for example, to apply colour settings to the colour and fill aesthetics at
the same time, via [aesthetics = c("colour", "fill")].}
}
\description{
scale_fill_dhsc_d
Use the DHSC discrete fill scale.
}
