% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{scale_fill_dhsc_red_green_c}
\alias{scale_fill_dhsc_red_green_c}
\title{scale_fill_dhsc_red_green_c}
\usage{
scale_fill_dhsc_red_green_c()
}
\description{
Use a continuous fill scale from DHSC green to red.
}
