% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_series.r
\name{sim_series}
\alias{sim_series}
\title{sim_series}
\usage{
sim_series(N = 1)
}
\arguments{
\item{N}{The number of rows to simulate}
}
\value{
A tibble of simulated data.
}
\description{
Generates dummy data for experimenting with plots. User specifies
the number of rows to generate and receives a tibble with three columns. A
character vector with unique values, the row number, and the row number squared.
}
\examples{
sim_series(5)
}
