% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dhsc.r
\name{theme_dhsc}
\alias{theme_dhsc}
\title{Theme DHSC for ggplot}
\usage{
theme_dhsc(base_size = 12)
}
\arguments{
\item{base_size}{Base font size, given in pts.}
}
\description{
Theme DHSC for ggplot
}
\examples{

 ggplot2::ggplot(dplyr::tibble(
 y = c(2, 4),
 x = c("A", "B")),
 ggplot2::aes(x = x, y = y, fill = x)) +
 ggplot2::geom_col() +
 theme_dhsc()

}
