% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc_functions.R
\name{zero_y_padding}
\alias{zero_y_padding}
\title{zero_y_padding}
\usage{
zero_y_padding()
}
\description{
By default, ggplot introduces some padding around the data to
ensure that they are placed some distance away from the axes. This function
is a shortcut to remove that padding so that if the data start at 0, the y
axis intercept will also be zero.
}
